requirements.txt
pytest
playwright
python-dotenv
aiohttp